grai-integrations/source-fivetran/pyproject.toml
[tool.poetry]
name = "grai_source_fivetran"
version = "0.0.1"
description = ""
authors = ["Ian Eaves <ian@grai.io>"]
license = "Elastic-2.0"
packages = [
    { include = "grai_source_fivetran", from = "src" },
]
[tool.poetry.dependencies]
python = "^3.9.13"
pydantic = "^1.9.1"
grai-client = "^0.1.3"
multimethod = "^1.8"
fivetran = "^0.7.0"
requests = "^2.28.1"
python-dotenv = "^0.21.1"
grai-schemas = "^0.1.9"

[tool.poetry.group.dev.dependencies]
black = "^22.6.0"
mypy = "^0.971"
isort = "^5.12.0"
pytest = "^7.2.0"
pre-commit = "^3.0.4"
types-requests = "^2.28.11.7"
fastapi = "^0.88.0"
fastapi-code-generator = "^0.3.6"


[tool.isort]
profile = "black"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
